requirements.txt
# Web framework
fastapi==0.135.3
uvicorn==0.43.0

# LLM orchestration
langchain==1.2.15
langchain-core==1.2.26
langchain-openai==1.1.12
langchain-ollama==1.0.1
langchain-huggingface==1.2.1
langchain-anthropic==1.4.0
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langchain-experimental==0.4.1

# Vector store
qdrant-client==1.17.1

# Embeddings & reranking
sentence-transformers==5.3.0

# BM25 search
rank-bm25==0.2.2

# PDF parsing
PyMuPDF==1.27.2.2

# Evaluation
ragas==0.4.3

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1

# Frontend
streamlit==1.56.0
requests==2.33.1

# Utilities
python-dotenv==1.2.2
pydantic==2.12.5
